ai_engine/requirements.txt
ultralytics
torch
opencv-python
requestsv 